An icon for the name: a latchkey, drawn rather than stored
The old icon was two nodes and the hop between them — a picture of the
MECHANISM. This one is a latchkey on a painted door: the key to your own front
door, which is what the new name means and what the app actually is. You are
not connecting to a service, you are letting yourself into your own house from
outside.

`design/latchkey-icon.swift` is the source and the PNGs are its output;
CoreGraphics only, no AppKit and no SVG rasteriser, so it runs from any shell
with Xcode. Three appearances, which the old catalog did not have: light, a
SUPPLIED dark (left alone iOS dims the light icon, and dimmed brass goes muddy
— M8.1 had already learned that), and a tinted grayscale-on-transparency
foreground, which is the documented input for that slot.

It renders a contact sheet next to the icons and I looked at it: true-pixel
rows at 29, 40, 60, 76, 120 and 180 on light and dark, plus the light icon on a
dark background. It holds at 29x29, which is the size that decides whether an
icon works — nothing is thinner than about 3 px there, and small sizes are
deliberately NOT drawn by hand, because Xcode derives them all from the 1024
and hand-drawing a 29 px variant would check an icon the phone never shows.

Verified with `xcrun actool` compiling the catalog standalone: no errors, no
warnings, derived sizes and Assets.car produced. Worth doing separately because
a full build was not available at the time — and because changing an asset
catalog under a running test tier could otherwise have failed a suite for a
reason that had nothing to do with it.

Two reservations, written into design/README.md rather than glossed: the glyph
is closer to a generic key than to a specifically latchkey silhouette, and its
bit reads a little like a letter E at large sizes. Both are shape problems
fixable in the generator alone. Committed as a working first version, not a
finished one.

The superseded generator and its PNG are removed, and the two places that
pointed at them now point here.

+++ design/latchkey-icon.swift
@@ -381,8 +381,8 @@ func sheet() -> CGImage {
     let big = 320
     place(masked(light, pixels: big), x: 60, y: 50, size: CGFloat(big))
     place(masked(tintedPreview(pixels: Int(side)), pixels: big), x: 400, y: 50, size: CGFloat(big))
-    place(masked(dark, pixels: big), x: W / 2 + 60, y: 50, size: CGFloat(big))
-    place(masked(light, pixels: big), x: W / 2 + 400, y: 50, size: CGFloat(big))
+    place(masked(dark, pixels: big), x: CGFloat(W) / 2 + 60, y: 50, size: CGFloat(big))
+    place(masked(light, pixels: big), x: CGFloat(W) / 2 + 400, y: 50, size: CGFloat(big))
 
     // Row 2: true pixel sizes, light on light and dark on dark, plus the light
     // icon on dark (a light-appearance user with a dark wallpaper).
